chore: split sandbox tools into separate package
