+++ packages/sandbox/README.md
@@ -1,0 +1,3 @@
+# @holoom/authority
+
+Helpers for creating and connecting to a holochain sandbox with holoom installed.
