+++ packages/sandbox/package.json
@@ -1,0 +1,48 @@
+{
+  "name": "@holoom/sandbox",
+  "version": "0.1.0-dev.13",
+  "description": "Helpers for starting a holochain sandbox with holoom installed",
+  "type": "module",
+  "license": "MIT",
+  "types": "dist/types/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/types/index.d.ts",
+      "node": {
+        "require": "./dist/index.node.cjs",
+        "default": "./dist/index.node.js"
+      }
+    }
+  },
+  "files": [
+    "dist/",
+    "src/",
+    "LICENSE",
+    "README.md"
+  ],
+  "repository": "https://github.com/holochain-open-dev/holoom",
+  "homepage": "https://github.com/holochain-open-dev/holoom/tree/main/packages/sandbox",
+  "bugs": {
+    "url": "https://github.com/holochain-open-dev/holoom.git/issues"
+  },
+  "scripts": {
+    "build": "rimraf dist && npm run build:node",
+    "build:node": "rollup -c rollup.node.config.ts --configPlugin typescript"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "devDependencies": {
+    "@rollup/plugin-alias": "^5.1.0",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/ws": "^8.5.10",
+    "rimraf": "^5.0.5",
+    "rollup": "^4.12.0",
+    "rollup-plugin-cleanup": "^3.2.1"
+  },
+  "dependencies": {
+    "@holochain/client": "^0.18.0-dev",
+    "yaml": "^2.5.1"
+  }
+}
